+++ calcGrad.jl
@@ -1,6 +1,6 @@
 
 
-@everywhere function nodesGradientReconstructionFastSA(beginCell::Int64, endCell::Int64, testMesh::mesh2d_shared, 
+@everywhere function nodesGradientReconstructionFastSA(beginCell::Int32, endCell::Int32, testMesh::mesh2d_shared, 
   scalarField::SharedArray{Float64,1}, gradX::SharedArray{Float64,1}, gradY::SharedArray{Float64,1}, gradMag::SharedArray{Float64,1})
 
 
@@ -18,13 +18,13 @@
     ny = zeros(Float64,4);
 
 
-      numNodesInCell::Int64 = testMesh.mesh_connectivity[C,3]; ## CMatrix mesh_connectivity - first index == 1
+      numNodesInCell::Int32 = testMesh.mesh_connectivity[C,3]; ## CMatrix mesh_connectivity - first index == 1
 	  
 	  areaCell::Float64 = testMesh.cell_areas[C]; 
 	  
-	  T1::Int64 = testMesh.mesh_connectivity[C,4];
-	  T2::Int64 = testMesh.mesh_connectivity[C,5];
-	  T3::Int64 = testMesh.mesh_connectivity[C,6];
+	  T1::Int32 = testMesh.mesh_connectivity[C,4];
+	  T2::Int32 = testMesh.mesh_connectivity[C,5];
+	  T3::Int32 = testMesh.mesh_connectivity[C,6];
 	  
 	  side[1] = testMesh.cell_edges_length[C,1];
 	  side[2] = testMesh.cell_edges_length[C,2];
@@ -54,7 +54,7 @@
 
 	  if (numNodesInCell == 4)
 	  
-	    T4::Int64 = testMesh.mesh_connectivity[C,7];
+	    T4::Int32 = testMesh.mesh_connectivity[C,7];
 		side[4] = testMesh.cell_edges_length[C,4];
 		nx[4] = testMesh.cell_edges_Nx[C,4];
 		ny[4] = testMesh.cell_edges_Nx[C,4];
@@ -111,7 +111,7 @@ end
 # end
 
 
-@everywhere function nodesGradientReconstructionFastPerThread(beginCell::Int64, endCell::Int64, testMesh::mesh2d_shared, 
+@everywhere function nodesGradientReconstructionFastPerThread(beginCell::Int32, endCell::Int32, testMesh::mesh2d_shared, 
   scalarField::SharedArray{Float64,1}, gradX::SharedArray{Float64,1}, gradY::SharedArray{Float64,1})
 
 
@@ -129,13 +129,13 @@ end
     ny = zeros(Float64,4);
 
 
-      numNodesInCell::Int64 = testMesh.mesh_connectivity[C,3]; ## CMatrix mesh_connectivity - first index == 1
+      numNodesInCell::Int32 = testMesh.mesh_connectivity[C,3]; ## CMatrix mesh_connectivity - first index == 1
 	  
 	  areaCell::Float64 = testMesh.cell_areas[C]; 
 	  
-	  T1::Int64 = testMesh.mesh_connectivity[C,4];
-	  T2::Int64 = testMesh.mesh_connectivity[C,5];
-	  T3::Int64 = testMesh.mesh_connectivity[C,6];
+	  T1::Int32 = testMesh.mesh_connectivity[C,4];
+	  T2::Int32 = testMesh.mesh_connectivity[C,5];
+	  T3::Int32 = testMesh.mesh_connectivity[C,6];
 	  
 	  side[1] = testMesh.cell_edges_length[C,1];
 	  side[2] = testMesh.cell_edges_length[C,2];
@@ -165,7 +165,7 @@ end
 
 	  if (numNodesInCell == 4)
 	  
-	    T4::Int64 = testMesh.mesh_connectivity[C,7];
+	    T4::Int32 = testMesh.mesh_connectivity[C,7];
 		side[4] = testMesh.cell_edges_length[C,4];
 		nx[4] = testMesh.cell_edges_Nx[C,4];
 		ny[4] = testMesh.cell_edges_Nx[C,4];
@@ -204,7 +204,7 @@ end
 
 
 
-@everywhere function nodesGradientReconstructionUconsFastSA(beginCell::Int64, endCell::Int64, testMesh::mesh2d_shared, 
+@everywhere function nodesGradientReconstructionUconsFastSA(beginCell::Int32, endCell::Int32, testMesh::mesh2d_shared, 
   ucons::SharedArray{Float64,2}, viscousFields2dX::viscousFields2d_shared)
 
 
@@ -225,13 +225,13 @@ end
 		ny = zeros(Float64,4);
 
 
-		  numNodesInCell::Int64 = testMesh.mesh_connectivity[C,3]; ## CMatrix mesh_connectivity - first index == 1
+		  numNodesInCell::Int32 = testMesh.mesh_connectivity[C,3]; ## CMatrix mesh_connectivity - first index == 1
 		  
 		  areaCell::Float64 = testMesh.cell_areas[C]; 
 		  
-		  T1::Int64 = testMesh.mesh_connectivity[C,4];
-		  T2::Int64 = testMesh.mesh_connectivity[C,5];
-		  T3::Int64 = testMesh.mesh_connectivity[C,6];
+		  T1::Int32 = testMesh.mesh_connectivity[C,4];
+		  T2::Int32 = testMesh.mesh_connectivity[C,5];
+		  T3::Int32 = testMesh.mesh_connectivity[C,6];
 		  
 		  side[1] = testMesh.cell_edges_length[C,1];
 		  side[2] = testMesh.cell_edges_length[C,2];
@@ -261,7 +261,7 @@ end
 
 		  if (numNodesInCell == 4)
 		  
-			T4::Int64 = testMesh.mesh_connectivity[C,7];
+			T4::Int32 = testMesh.mesh_connectivity[C,7];
 			side[4] = testMesh.cell_edges_length[C,4];
 			nx[4] = testMesh.cell_edges_Nx[C,4];
 			ny[4] = testMesh.cell_edges_Nx[C,4];
